+++ modulo2/projeto-whatslab/projeto-whatslab/package.json
@@ -8,7 +8,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^3.0.1",
     "styled-component": "^2.8.0",
     "styled-components": "^5.3.3",
     "web-vitals": "^2.1.3"
